credits done

+++ base funzionante/SpringBoot-Angular7-Online-Shopping-Store-master/backend/src/main/java/com/project/Toshop/service/UserService.java
@@ -24,7 +24,7 @@ public interface UserService {
 
     Client updateClient(Client client);
 
-    Client updateCredits(BigDecimal productPrice, Long idUtente);
+    Client updateCredits(int productPrice, Long idUtente);
 
     Supplier updateSupplier(Supplier supplier);
